refactor(council): relocate DebateEvent vocabulary to value_objects
Per the council pilot reviewer nit. Debate events are immutable records
with no identity beyond their data, which is the value-object
definition. Keeps Argument and SubtaskBrief in domain/entities/council
(they are the debate inputs/outputs, with SubtaskBrief carrying an id).

- New: domain/value_objects/council_events.py with DebateStarted /
  ArgumentSubmitted / DecisionResolved / DebateAbandoned / DebateEvent /
  DebateEventAdapter
- Slim: domain/entities/council.py now holds only Argument and
  SubtaskBrief
- Update imports in application use case, port, infra adapter, fakes,
  and unit + integration tests

No behavior change. 11 council unit tests pass; ruff clean.
